# Conflicts:
#	HODLER_Multiplatform_Win_And_iOS_Linux/HODLER_Wallet_Desktop.dpr
#	HODLER_Multiplatform_Win_And_iOS_Linux/HODLER_Wallet_Desktop.dproj
#	HODLER_Multiplatform_Win_And_iOS_Linux/HODLER_Wallet_Desktop.dproj.local
#	HODLER_Multiplatform_Win_And_iOS_Linux/HODLER_Wallet_Desktop.dres
#	HODLER_Multiplatform_Win_And_iOS_Linux/HODLER_Wallet_Desktop.identcache
#	HODLER_Multiplatform_Win_And_iOS_Linux/HODLER_Wallet_Desktop.stat
#	HODLER_Multiplatform_Win_And_iOS_Linux/HODLER_Wallet_DesktopResource.rc
#	HODLER_Multiplatform_Win_And_iOS_Linux/HODLER_Wallet_Mobile.identcache
#	HODLER_Multiplatform_Win_And_iOS_Linux/HODLER_Wallet_Mobile.stat
#	HODLER_Multiplatform_Win_And_iOS_Linux/homeAndroid/uHome.SmXhdpiPh.fmx
#	HODLER_Multiplatform_Win_And_iOS_Linux/homeAndroid/uHome.Surface.fmx
#	HODLER_Multiplatform_Win_And_iOS_Linux/homeAndroid/uHome.Windows.fmx
#	HODLER_Multiplatform_Win_And_iOS_Linux/homeAndroid/uHome.fmx
#	HODLER_Multiplatform_Win_And_iOS_Linux/homeAndroid/uHome.iPhone55in.fmx
#	HODLER_Multiplatform_Win_And_iOS_Linux/homeAndroid/uHome.pas
#	HODLER_Multiplatform_Win_And_iOS_Linux/homeWindows/uHome.LgXhdpiPh.fmx
#	HODLER_Multiplatform_Win_And_iOS_Linux/homeWindows/uHome.SmXhdpiPh.fmx
#	HODLER_Multiplatform_Win_And_iOS_Linux/homeWindows/uHome.Surface.fmx
#	HODLER_Multiplatform_Win_And_iOS_Linux/homeWindows/uHome.Windows.fmx
#	HODLER_Multiplatform_Win_And_iOS_Linux/homeWindows/uHome.fmx
#	HODLER_Multiplatform_Win_And_iOS_Linux/homeWindows/uHome.iPhone55in.fmx
#	HODLER_Multiplatform_Win_And_iOS_Linux/homeWindows/uHome.pas

+++ HODLER_Multiplatform_Win_And_iOS_Linux/additionalUnits/misc.pas
@@ -316,7 +316,7 @@ procedure refreshCurrencyValue();
 procedure updateBalanceLabels();
 
 Function StrToQRBitmap(Str: AnsiString; pixelSize: integer = 6): TBitmap;
-procedure shareFile(path: AnsiString);
+procedure shareFile(path: AnsiString;deleteSourceFile:Boolean=true);
 procedure synchronizeCurrencyValue();
 procedure LoadCurrencyFiatFromFile();
 function bitcoinCashAddressToCashAddress(address: AnsiString): AnsiString;
@@ -389,7 +389,7 @@ var
 
 implementation
 
-uses Bitcoin, uHome, base58, Ethereum, coinData, strutils, secp256k1
+uses Bitcoin, uHome, base58, Ethereum, coinData, strutils, secp256k1 ,AccountRelated
 {$IFDEF ANDROID}
 {$ELSE}
 {$ENDIF};
@@ -1251,7 +1251,9 @@ begin
       Tthread.Synchronize(nil,
         procedure
         begin
-          frmhome.FormShow(nil);
+           frmhome.FormShow(nil);
+        AccountRelated.LoadCurrentAccount(name);
+
         end);
 
     end);
@@ -1663,7 +1665,7 @@ begin
   result.scriptHash := reverseHexOrder(GetSHA256FromHex((result.outputScript)));
 end;
 
-procedure shareFile(path: AnsiString);
+procedure shareFile(path: AnsiString;deleteSourceFile:Boolean=true);
 var
   i: integer;
 {$IFDEF ANDROID}
@@ -1719,7 +1721,7 @@ begin
   if saveDialog.Execute then
   begin
     TFile.Copy(path, saveDialog.FileName);
-    DeleteFile(path);
+    if deleteSourceFile then DeleteFile(path);
 
   end;
   saveDialog.Free;
@@ -1792,7 +1794,7 @@ begin
       if QRCodeBitmap.data <> nil then
       begin
         bmp.Free;
-        bmp := BitmapDataToScaledBitmap(QRCodeBitmap, 6);
+        bmp := BitmapDataToScaledBitmap(QRCodeBitmap, pixelSize);
         bmp.Unmap(QRCodeBitmap);
       end;
     end;
